Adds support for return and initial pthread support
Support for pthread_create() and pthread_join() is added. When calling
pthread_create(), a new thread is started by finding an empty process slot
and initializing the stack, registers and PC of that process slot. The PC
is set to the start of the function specified in pthread_create(). Note
that arbitrary function pointers are NOT supported. This could change.
When calling pthread_join(), a transition is only reported when the thread
to join has stopped and put its result in the thread result list.